backend/requirements.txt
fastapi
uvicorn
yfinance
pandas_ta
finnhub-python
pandas
numpy
python-multipart